Model as layer.
