+++ juicer/util/template_util.py
@@ -279,7 +279,7 @@ def get_random_interval(field):
     if field is not None:
         if isinstance(field, basestring) and field.strip():
             field = re.sub(r"\{|\[|\}|\]||\(|\)|\s+", "", field)
-            field = field.split(':')
+            field = field.split(',')
             if len(field) == 2:
                 return ', '.join(field)
     return False
